Add play button in index.html which uses playGame from script.js
Edit index.html to defer the script.js execution and add play button which onClick() runs playGame() from the script.
Add a new game counter to count number of rounds to be played in a game.
User choice prompt includes an option to abort the game, and displays current user's and machine's score along with number of rounds left to play.
The function playRound() returns the same result for draw and invalid user choice outcomes. Also, simultaneously utilises rounds counter.
The function playGame() loops through playRound() function for the number of rounds to be played by conditionally using the rounds counter and logging round winners, and the game winner once all the round results are decisively over.